Extract Formula component from variable
