+++ src/components/variable.jsx
@@ -1,14 +1,13 @@
 import DocumentTitle from "react-document-title"
 import { FormattedMessage, FormattedDate, FormattedHTMLMessage } from "react-intl"
 import React, { PropTypes } from "react"
-import { Link } from "react-router"
-import { keys, flatten, pipe, map, is } from "ramda"
+import { keys } from "ramda"
 
 import config from "../config"
 import * as AppPropTypes from "../app-prop-types"
 import ExternalLink from "./external-link"
+import Formula from "./formula"
 import getDayBefore from "../periods"
-import Highlight from 'react-highlight'
 
 const Variable = React.createClass({
   propTypes: {
@@ -192,133 +191,20 @@ const Variable = React.createClass({
             const startDate = (date != "0001-01-01") && date
             const stopDate = startDates[dateIndex - 1] && getDayBefore(startDates[dateIndex - 1])
             return formulas[date] && (
-              <div key={ date } >
-                {startDate && (! stopDate) &&
-                  <h3>À partir du <FormattedDate value={ startDate } />&nbsp;:</h3>
-                }
-                {stopDate && (! startDate) &&
-                  <h3>Jusqu'au <FormattedDate value={ stopDate } />&nbsp;:</h3>
-                }
-                {startDate && stopDate &&
-                  <h3>Du <FormattedDate value={ startDate } /> au <FormattedDate value={ stopDate } />&nbsp;:</h3>
-                }
-                <Highlight className="python"> { this.renderLinkedFormula(formulas[date].content) } </Highlight>
-              <p>
-                 <a href={ formulas[date].source }  target="_blank"><FormattedMessage id="editThisFormula"/></a>
-              </p>
-              </div>
+              <Formula
+                key={startDate}
+                start-date={startDate}
+                stop-date={stopDate}
+                content={formulas[date].content}
+                source={formulas[date].source}
+                variables={this.props.variables}
+                parameters={this.props.parameters}
+                />
             )
           })
         }
       </div>
     )
-  },
-  link(variable) {
-    return <Link key={ variable + Math.random() } to={ variable }
-      data-toggle="popover" title={ this.props.variables[variable].description }>{ variable }</Link>
-  },
-  linkParam(parameter, linkText) {
-    return <Link key={parameter + Math.random()} to={parameter}
-      data-toggle="popover" title={this.props.parameters[parameter].description}>{ linkText }</Link>
-  },
-  isVariable(substring) {
-    // Ignore every text that isn't a single word like a variable must be:
-    return (! substring.includes(" ") && this.props.variables[substring])
-  },
-
-  splitAndLinkVariables(text, separator) {
-    // Only split strings, as trying to split JSX Links would raise an error
-    if (! is(String, text)) {
-      return text
-    }
-    const splits = text.split(separator)
-    return splits.map((substring, index) => {
-      if (this.isVariable(substring)) {
-        substring = [this.link(substring), separator]  // Concatenate JSX with a string (+ doesn't work).
-      } else {
-        substring = index < splits.length - 1 ? substring + separator : substring
-      }
-      return substring
-    })
-  },
-
-  splitAndLinkParams(text) {
-    const recordedParamNodes = []
-    const recordParamNode = (nodeVariableName, path) => {
-      // Matches "P_2 = P.af" or "P.af"
-      const splitRegex = new RegExp(`((?:[\\S]*[\\s]*[=][\\s]*)?${nodeVariableName}\\.[0-9a-zA-Z\\_\\.]*)`)
-      const matchRegex = new RegExp(`(?:([\\S]*)([\\s]*[=][\\s]*))?(${nodeVariableName}\\.)([0-9a-zA-Z\\_\\.]*)`)
-      recordedParamNodes.push({
-        'nodeVariableName': nodeVariableName,
-        'parameterPath': path,
-        'splitRegex': splitRegex,
-        'matchRegex': matchRegex
-      })
-    }
-    // Matches "P = parameters(period).af"
-    const splits = text.split(/([\S]*[\s]*[=][\s]* parameters\([\S]*\)\.?[\S]*)/)
-    // We go through each substring and find parameters(x=parameter(x).x.x) and check if they are a node or an actual parameter.
-    // If it's a node, it records the node and the variable associated. Else, it returns the link.
-    return splits.map((substring, index) => {
-      // This checks if the split actually happened
-      if (index % 2 == 1) {
-        const parameterCall = substring.match(/([\S]*)([\s]*[=][\s]*parameters\([\S]*\)\.?)([\S]*)/)
-        const parameterPath = parameterCall[3]
-        const nodeVariableName = parameterCall[1]
-          if (this.props.parameters[parameterPath]) {
-            return [nodeVariableName, parameterCall[2], this.linkParam(parameterPath, parameterPath)]  // Concatenate JSX with a string (+ doesn't work).
-          } else {
-            recordParamNode(nodeVariableName,parameterPath)
-            return [substring]
-          }
-      } else {
-        let substrings = [substring]
-        for (let i = 0; i < recordedParamNodes.length; i++) {
-          const element = recordedParamNodes[i]
-          substrings = substrings.map(substring2 => {
-            return ! is(String, substring2)
-              ? substring2
-              : substring2
-                .split(element.splitRegex)
-                .map(substring3 => {
-                  const match = substring3.match(element.matchRegex)
-                  if ( ! match) {
-                    return substring3
-                  } else {
-                    const parameterPath = match[4]
-                    const nodePath = element.parameterPath ? `${element.parameterPath}.${parameterPath}` : parameterPath
-                    if (this.props.parameters[nodePath]) {
-                      const linkThisParam = this.linkParam(nodePath, parameterPath)
-                      return [match[1],match[2],match[3], linkThisParam]
-                    } else {
-                      const nodeVariableName = match[1]
-                      recordParamNode(nodeVariableName,nodePath)
-                      return [substring3]
-                    }
-                  }
-                })
-          })
-          substrings = flatten(substrings)
-        }
-        return substrings
-      }
-    })
-  },
-  // Change every OpenFisca parameter and variable in the formula by a link to the variable page:
-  renderLinkedFormula(formula) {
-    // Split on double quotes first (preventing collision with Link):
-    return pipe(
-      this.splitAndLinkParams,
-      flatten,
-      map((substring) => {
-        return this.splitAndLinkVariables(substring, `"`)
-      }),
-      flatten,
-      map((substring) => {
-        return this.splitAndLinkVariables(substring, `'`)
-      }),
-      flatten,
-      )(formula)
   },
 })
