+++ tests/unit/formulas.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Variable from '../../src/components/variable'
+import Formula from '../../src/components/formula'
 import { shallow } from 'enzyme';
 import {flatten} from 'ramda'
 import {Link} from 'react-router'
@@ -24,45 +24,25 @@ const parameters = {
 }
 
 const variable = {id: 'rsa'}
-const component = shallow(<Variable variable={variable} variables={variables} parameters={parameters} />).instance()
 
-// describe('Parameter checking', function(){
-//     it ('should return true when given a parameter', function(){
-//         const substring = 'parameters(period.this_year.first_month).prestations.prestations_familiales.af.bmaf '
-//         const output = component.isParameterCall(substring)
-//         output.should.be.true()
-//     })
-//     it ('should return false when given a random string', function(){
-//         const substring = "scolarite_i = famille.members('scolarite',"
-//         const output = component.isParameterCall(substring)
-//         output.should.be.false()
-//     })
-//     it ('should return true when parameter node is a leaf', function(){
-//         const substring = 'P = parameters(period.this_year.first_month).prestations.prestations_familiales.af.bmaf'
-//         const output = component.isParameterLeaf(substring)
-//         output.should.be.ok()
-//     })
-//     it('should return false if parameter is a node', function(){
-//         const substring = 'P = parameters(period).bourses_education.bourse_college'
-//         const output = component.isParameterLeaf(substring)
-//         should(output).not.be.ok()
-//     })
-// })
+const component = shallow(<Formula variable={variable} variables={variables} parameters={parameters} content="" />).instance()
 
 function splitAndLinkParams(formula) {
     return shallow(<div>{component.splitAndLinkParams(formula)}</div>)
 }
+
+
 describe('Add links to parameters', function(){
     it ('should return one link when there is one leaf', function(){
-        const formula = fs.readFileSync(path.join(__dirname, "formula2.txt")).toString()
-        const output = splitAndLinkParams(formula)
+        const formula_content = fs.readFileSync(path.join(__dirname, "formula2.txt")).toString()
+        const output = splitAndLinkParams(formula_content)
         const links = output.find(Link)
         links.should.have.length(1)
         links.get(0).props.to.should.equal('prestations.prestations_familiales.af.bmaf')
     })
     it ('should return a link for each parameter present', function(){
-        const formula = fs.readFileSync(path.join(__dirname, "formula1.txt")).toString()
-        const output = splitAndLinkParams(formula)
+        const formula_content = fs.readFileSync(path.join(__dirname, "formula1.txt")).toString()
+        const output = splitAndLinkParams(formula_content)
         const links = output.find(Link)
         links.should.have.length(4)
         links.get(0).props.to.should.equal('prestations.prestations_familiales.af.bmaf')
@@ -71,14 +51,13 @@ describe('Add links to parameters', function(){
         links.get(3).props.to.should.equal('bourses_education.bourse_college.montant_taux_1')
     })
     it ('should return a link when embeded in several nodes', function(){
-        const formula = fs.readFileSync(path.join(__dirname, "formula4.txt")).toString()
-        const output = splitAndLinkParams(formula)
+        const formula_content = fs.readFileSync(path.join(__dirname, "formula4.txt")).toString()
+        const output = splitAndLinkParams(formula_content)
         const links = output.find(Link)
         links.should.have.length(1)
         links.get(0).props.to.should.equal('prestations.prestations_familiales.af.bmaf')
     })
 })
-
 
 function renderLinkedFormula(formula) {
     return shallow(<div>{component.renderLinkedFormula(formula)}</div>)
@@ -87,8 +66,8 @@ function renderLinkedFormula(formula) {
 describe('Add links to the whole formula', function(){
 
     it ('should return 2 links when there is very one parameter and one variable', function(){
-        const formula = fs.readFileSync(path.join(__dirname, "formula3.txt")).toString()
-        const output = renderLinkedFormula(formula)
+        const formula_content = fs.readFileSync(path.join(__dirname, "formula3.txt")).toString()
+        const output = renderLinkedFormula(formula_content)
         const links = output.find(Link)
         links.should.have.length(2)
         links.get(0).props.to.should.equal('rsa_eligible')
@@ -96,16 +75,16 @@ describe('Add links to the whole formula', function(){
 
     })
     it ('should return 2 links when there is one parameter and one variable and a node', function(){
-        const formula = fs.readFileSync(path.join(__dirname, "formula2.txt")).toString()
-        const output = renderLinkedFormula(formula)
+        const formula_content = fs.readFileSync(path.join(__dirname, "formula2.txt")).toString()
+        const output = renderLinkedFormula(formula_content)
         const links = output.find(Link)
         links.should.have.length(2)
         links.get(0).props.to.should.equal('rsa_eligible')
         links.get(1).props.to.should.equal('prestations.prestations_familiales.af.bmaf')
     })
     it ('should return 6 links', function(){
-        const formula = fs.readFileSync(path.join(__dirname, "formula1.txt")).toString()
-        const output = renderLinkedFormula(formula)
+        const formula_content = fs.readFileSync(path.join(__dirname, "formula1.txt")).toString()
+        const output = renderLinkedFormula(formula_content)
         const links = output.find(Link)
         links.should.have.length(6)
         links.get(0).props.to.should.equal('prestations.prestations_familiales.af.bmaf')
@@ -120,23 +99,23 @@ describe('Add links to the whole formula', function(){
 
 describe('Link rendering in variables', function() {
     it('should add links to a variables with double quotes', function() {
-        const formula = 'simulation.calculate("rsa_base_ressource")'
-        const output = renderLinkedFormula(formula)
+        const formula_content = 'simulation.calculate("rsa_base_ressource")'
+        const output = renderLinkedFormula(formula_content)
         const links = output.find(Link)
         links.should.have.length(1)
         links.get(0).props.to.should.equal('rsa_base_ressource')
     });
     it('should add links to a variables with simple quotes', function() {
-        const formula = 'simulation.calculate(\'rsa_base_ressource\')'
-        const output = renderLinkedFormula(formula)
+        const formula_content = 'simulation.calculate(\'rsa_base_ressource\')'
+        const output = renderLinkedFormula(formula_content)
         const links = output.find(Link)
         links.should.have.length(1)
         links.get(0).props.to.should.equal('rsa_base_ressource')
     });
 
     it('should handle simple quotes and double quotes mixed', function() {
-        const formula = 'simulation.calculate(\'rsa_base_ressource\') simulation.calculate("rsa_fictif") simulation.calculate(\'rsa_eligible\')'
-        const output = renderLinkedFormula(formula)
+        const formula_content = 'simulation.calculate(\'rsa_base_ressource\') simulation.calculate("rsa_fictif") simulation.calculate(\'rsa_eligible\')'
+        const output = renderLinkedFormula(formula_content)
         const links = output.find(Link)
         links.should.have.length(3)
         links.get(0).props.to.should.equal('rsa_base_ressource');
@@ -145,8 +124,8 @@ describe('Link rendering in variables', function() {
     });
 
     it('should be robust to quotes in comments', function() {
-        const formula = ' # This a a comment with a quote \' simulation.calculate(\'rsa_base_ressource\') simulation.calculate(\'rsa_eligible\')'
-        const output = renderLinkedFormula(formula)
+        const formula_content = ' # This a a comment with a quote \' simulation.calculate(\'rsa_base_ressource\') simulation.calculate(\'rsa_eligible\')'
+        const output = renderLinkedFormula(formula_content)
         const links = output.find(Link)
         links.should.have.length(2)
         links.get(0).props.to.should.equal('rsa_base_ressource');
